Fix a potential memory leak
In the case where looking for the action group would return a nullptr,
an allocation would be made. Right after, there is a check for the
number of ports, and if it 0, the function doesn't do anything more.
At this point, the allocated action_group is then lost.

This commit changes slightly the logic by checking for ports firsts,
and only if there is some, then check if an allocation is required or
not. If there is at least one port, one is sure to enter the loop, in
which the action group is given to some Qt class that handles the
resource lifetime.